Fix glitch where data would get overwritten when using several sport_keys
Refactor the code to store the filtered data for each sport in a separate dictionary. This change ensures that data for different sports is not overwritten .The `sport_filtered_data` dictionary is initialized using a dictionary comprehension, where each sport key is associated with an empty list. The filtered data for each sport is then appended to the corresponding list within the dictionary. This change allows for accurate data accumulation and processing when creating the final DataFrame.